Format manuals and torture tests.
'webman.tex' is in a TeX Live package, but the other documents are in
the TeX Live source tree.

TODO: 'trapman' covers only Metafont, but in the TL source tree are
input and output for Metafont and MetaPost.  Some of the 'trap' files
have to be prefixed with 'mf*' (manually on the first try).