Work around issue cwebbin#32.
Setting '\gtitle' right before '\input pdfwebtocfront' fixes the group
title on the first page after the (shifted) table-of-contents page.
This _should_ be 'Introduction' in most cases, but Metafont, TeX,
pdfTeX, and XeTeX use the more comple 'Part 1: Introduction' in their
preamble.  The first '\N' should do the same, but starting with the
second page, the automatic '\mark' mechanism kicks in.