Compile WEAVE.WEB with Free Pascal.
Now that I got the hang of the matter, it's quite simple to adapt the
Pascal-H code in WEAVE.WEB to the conventions of FPC in one fell swoop.

Note that FPC -- correctly -- issues a warning for module 249, i.e., the
implementation of the 'unbucket' procedure: Given the definitions of
'sort_ptr' (==scrap_ptr) and 'max_sorts' (==max_scraps), it's virtually
impossible that the condition 'sort_ptr>max_sorts' would ever be 'true',
because a decent Pascal compiler -- like FPC -- would not allow an index
variable like 'sort_ptr: 0..max_sorts' to overrun its limits.