Help PDFTeX to produces valid output.
By default, 'pwebmac.tex' uses the final value of '\modno' (in general
the module number of the 'Index.') as the number of subentries of the
'NOS' node in the PDF outlines/bookmarks. This leads to warnings from
various online PDF validators.

Place the shell script 'countmodules' somewhere in your $PATH.

Invoke the shell script 'makeall' with options '--tex=pdftex --validpdf'
(shorter '-t pdftex -v').

Macro package 'pwebmac.tex' is set up to register when 'pdftex' is
invoked with the '--shell-escape' option and resets '\modno' to the
output of 'countmodules \jobname.tex' at the end of the 'pdftex' run.