Handle spreads with empty mini-index.
At least for TANGLE.WEB and WEAVE.WEB, the *.ref file created in the
first pass of TeX ends with a '!\d+' indicator line. This must be
shipped to the *.sref file in order to satisfy TeX in the second pass.

Now the whole TWILL/TeX/RefSort/TeX process runs smoothly, provided that
all '\none' entries in the TeX sources are resolved correctly.